+++ digital-ocean/container-linux/kubernetes/variables.tf
@@ -71,12 +71,6 @@ variable "asset_dir" {
   type        = "string"
 }
 
-variable "networking" {
-  description = "Choice of networking provider (flannel or calico)"
-  type        = "string"
-  default     = "flannel"
-}
-
 variable "pod_cidr" {
   description = "CIDR IPv4 range to assign Kubernetes pods"
   type        = "string"